Split tests between spc styles
RE #
