+++ tests/mercuryitc.py
@@ -1,8 +1,5 @@
-import contextlib
 import itertools
 import unittest
-from datetime import time
-from time import sleep
 
 from parameterized import parameterized
 
@@ -55,25 +52,9 @@ macros = {}
 macros.update({"TEMP_{}".format(key): val for key, val in enumerate(TEMP_CARDS, start=1)})
 macros.update({"PRESSURE_{}".format(key): val for key, val in enumerate(PRESSURE_CARDS, start=1)})
 macros.update({"LEVEL_{}".format(key): val for key, val in enumerate(LEVEL_CARDS, start=1)})
-macros["FLOW_SPC_PRESSURE_1"] = 1
-macros["FLOW_SPC_MIN_PRESSURE"] = SPC_MIN_PRESSURE
-macros["FLOW_SPC_TEMP_DEADBAND"] = SPC_TEMP_DEADBAND
-macros["FLOW_SPC_MAX_PRESSURE"] = SPC_MAX_PRESSURE
-macros["FLOW_SPC_OFFSET"] = SPC_OFFSET
-macros["FLOW_SPC_OFFSET_DURATION"] = SPC_OFFSET_DURATION
-macros["FLOW_SPC_GAIN"] = SPC_GAIN
 
 macros["CALIB_BASE_DIR"] = EPICS_TOP.replace("\\", "/")
 macros["CALIB_DIR"] = os.path.join("support", "mercuryitc", "master", "settings").replace("\\", "/")
-macros["FLOW_SPC_TABLE_FILE"] = "little_blue_cryostat.txt"
-
-macros["VTI_SPC_PRESSURE_1"] = 1
-macros["VTI_SPC_MIN_PRESSURE"] = 10
-macros["VTI_SPC_MAX_PRESSURE"] = 50
-macros["VTI_SPC_PRESSURE_CONSTANT"] = 5
-macros["VTI_SPC_TEMP_CUTOFF_POINT"] = 5
-macros["VTI_SPC_TEMP_SCALE"] = 5
-
 
 
 DEVICE_PREFIX = "MERCURY_01"
@@ -151,10 +132,7 @@ class MercuryTests(unittest.TestCase):
         self._lewis.backdoor_set_on_device("connected", True)
         self.ca = ChannelAccess(device_prefix=DEVICE_PREFIX, default_timeout=20)
         card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        self.ca.assert_setting_setpoint_sets_readback("OFF", readback_pv="{}:SPC".format(card_pv_prefix),
-                                                      expected_alarm=self.ca.Alarms.MAJOR)
         
-
     @parameterized.expand(parameterized_list(
         itertools.product(PID_PARAMS, PID_TEST_VALUES, TEMP_CARDS + PRESSURE_CARDS)))
     @skip_if_recsim("Lewis backdoor not available in recsim")
@@ -392,146 +370,6 @@ class MercuryTests(unittest.TestCase):
             "{}:RESISTANCE".format(get_card_pv_prefix(PRIMARY_TEMPERATURE_CHANNEL)), self.ca.Alarms.NONE
         )
 
-    def test_WHEN_auto_flow_set_THEN_pv_updates_and_states_are_set(self):
-        card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        self.ca.set_pv_value("{}:PID:AUTO:SP".format(card_pv_prefix), "OFF")
-        self.ca.set_pv_value("{}:FLOW:STAT:SP".format(card_pv_prefix), "Auto")
-        self.ca.set_pv_value("{}:HEATER:MODE:SP".format(card_pv_prefix), "Manual")
-        self.ca.set_pv_value("{}:FLOW:STAT:SP".format(pressure_card_pv_prefix), "Manual")
-
-        self.ca.assert_setting_setpoint_sets_readback(
-            "ON", set_point_pv="{}:SPC:SP".format(card_pv_prefix),
-            readback_pv="{}:SPC".format(card_pv_prefix))
-        self.ca.assert_that_pv_is("{}:PID:AUTO".format(card_pv_prefix), "ON")
-        self.ca.assert_that_pv_is("{}:FLOW:STAT".format(card_pv_prefix), "Manual")
-        self.ca.assert_that_pv_is("{}:HEATER:MODE".format(card_pv_prefix), "Auto")
-        self.ca.assert_that_pv_is("{}:FLOW:STAT".format(pressure_card_pv_prefix), "Auto")
-        self.ca.assert_that_pv_is("{}:SPC:SP".format(pressure_card_pv_prefix), "ON")
-
-    def test_WHEN_auto_flow_set_off_THEN_pv_updates_and_states_are_not_set(self):
-        card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        self.ca.set_pv_value("{}:PID:AUTO:SP".format(card_pv_prefix), "OFF")
-        self.ca.set_pv_value("{}:FLOW:STAT:SP".format(card_pv_prefix), "Auto")
-        self.ca.set_pv_value("{}:HEATER:MODE:SP".format(card_pv_prefix), "Manual")
-        self.ca.set_pv_value("{}:FLOW:STAT:SP".format(pressure_card_pv_prefix), "Manual")
-
-        self.ca.assert_setting_setpoint_sets_readback(
-            "OFF", set_point_pv="{}:SPC:SP".format(card_pv_prefix),
-            readback_pv="{}:SPC".format(card_pv_prefix), expected_alarm=self.ca.Alarms.MAJOR)
-        self.ca.assert_that_pv_is("{}:PID:AUTO".format(card_pv_prefix), "OFF")
-        self.ca.assert_that_pv_is("{}:FLOW:STAT".format(card_pv_prefix), "Auto")
-        self.ca.assert_that_pv_is("{}:HEATER:MODE".format(card_pv_prefix), "Manual")
-        self.ca.assert_that_pv_is("{}:FLOW:STAT".format(pressure_card_pv_prefix), "Manual")
-        self.ca.assert_that_pv_is("{}:SPC:SP".format(pressure_card_pv_prefix), "OFF")
-
-    def set_temp_reading_and_sp(self, reading, set_point, spc_state="On"):
-        """
-        Set the temperature in lewis and the set point on the first card
-        :param reading: The reading lewis will return
-        :param set_point: The set point to set on the device
-        :param spc_state: State to set SPC to (defaults to On)
-        """
-        card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        self.ca.set_pv_value("{}:SPC:SP".format(card_pv_prefix), spc_state)
-        self.ca.set_pv_value("{}:TEMP:SP".format(card_pv_prefix), set_point)
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", reading])
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_on_and_temp_lower_than_2_deadbands_THEN_pressure_set_to_minimum_pressure(self):
-        set_point = 10
-        reading = 10 - SPC_TEMP_DEADBAND * 2.1
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        self.set_temp_reading_and_sp(reading, set_point)
-
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP:RBV".format(pressure_card_pv_prefix), SPC_MIN_PRESSURE)
-
-    @parameterized.expand([(10, ), (1, ), (300, ), (12, ), (20,)])
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_on_and_temp_low_but_within_1_to_2_deadbands_THEN_pressure_set_to_pressure_for_setpoint_temp_and_does_not_ramp(self, set_point):
-        reading = set_point - SPC_TEMP_DEADBAND * 1.5
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        self.set_temp_reading_and_sp(reading, set_point)
-
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP:RBV".format(pressure_card_pv_prefix), pressure_for(set_point))
-        sleep(1.5)
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP:RBV".format(pressure_card_pv_prefix), pressure_for(set_point))
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_on_and_temp_at_setpoint_THEN_pressure_set_to_pressure_for_setpoint_temp_and_does_ramp_down(self):
-        set_point = 10
-        reading = set_point
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        self.set_temp_reading_and_sp(reading, set_point)
-
-        self.ca.assert_that_pv_is_number("{}:PRESSURE:SP".format(pressure_card_pv_prefix), pressure_for(set_point) + SPC_OFFSET, tolerance=SPC_OFFSET/4)  # should see number in ramp
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), pressure_for(set_point))  # final value
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_on_and_temp_above_setpoint_by_more_than_deadband_THEN_pressure_set_to_pressure_for_setpoint_temp_plus_gain_and_does_ramp(self):
-        diff = SPC_TEMP_DEADBAND * 1.1
-        set_point = 10
-        reading = set_point + diff
-        expected_pressure = pressure_for(set_point) + SPC_OFFSET + (abs(reading - set_point - SPC_TEMP_DEADBAND) * SPC_GAIN) ** 2
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        self.set_temp_reading_and_sp(reading, set_point)
-
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_pressure)  # final value
-        sleep(1.5)  # wait for possible ramp
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_pressure)  # final value
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_on_and_pressure_would_be_high_THEN_pressure_set_to_maximum_pressure(self):
-        diff = 1000
-        set_point = 10
-        reading = set_point + diff
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        self.set_temp_reading_and_sp(reading, set_point)
-
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), SPC_MAX_PRESSURE)  # final value
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_set_off_THEN_pressure_is_not_updated(self):
-        diff = 1000
-        set_point = 10
-        reading = set_point + diff
-
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        expected_value = -10
-        self.ca.set_pv_value("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_value)
-
-        self.set_temp_reading_and_sp(reading, set_point, "OFF")
-
-        self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_value)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_WHEN_auto_flow_on_but_error_in_temp_readback_THEN_pressure_is_not_updated(self):
-        set_point = 10
-
-        card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-        expected_value = -10
-        self.ca.set_pv_value("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_value)
-
-        with self._lewis.backdoor_simulate_disconnected_device():
-            # we have a lot of db records scanning in the ioc, they will start to fail but it may take a while
-            # (and this will be variable) before the record we choose gets processed and fails or hits
-            # the stream device lock timeout instead.
-            # So we need to wait for at least as long as stream device lock timeout to see an alarm raised`
-            self.ca.assert_that_pv_alarm_is("{}:TEMP:SP:RBV".format(card_pv_prefix), self.ca.Alarms.INVALID, timeout=60)
-
-            self.ca.set_pv_value("{}:SPC:SP".format(card_pv_prefix), "ON")
-            self.ca.set_pv_value("{}:TEMP:SP".format(card_pv_prefix), set_point)
-
-            self.ca.assert_that_pv_is("{}:PRESSURE:SP".format(pressure_card_pv_prefix), expected_value)
-        
-        self.ca.assert_that_pv_alarm_is("{}:TEMP:SP:RBV".format(card_pv_prefix), self.ca.Alarms.NONE, timeout=60)
-
     @parameterized.expand(parameterized_list(
         itertools.product(TEMP_CARDS + PRESSURE_CARDS)))
     @skip_if_recsim("Lewis backdoor not available in recsim")
@@ -580,313 +418,3 @@ class MercuryTests(unittest.TestCase):
         self.ca.set_pv_value("{}:HEATER:VOLT_LIMIT:SP".format(card_pv_prefix), new_limit)
 
         self.ca.assert_that_pv_is("{}:HEATER:VOLT_PRCNT".format(card_pv_prefix), new_percent)
-
-class MercuryVTISPCTests(unittest.TestCase):
-    """
-    Tests for the Mercury IOC VTI Software Pressure Control.
-    """
-    def setUp(self):
-        self._lewis, self._ioc = get_running_lewis_and_ioc("mercuryitc", DEVICE_PREFIX)
-        self._lewis.backdoor_set_on_device("connected", True)
-        self.ca = ChannelAccess(device_prefix=DEVICE_PREFIX, default_timeout=20)
-        self._change_spc_pressure_lookup_table("None.txt")
-
-    def test_WHEN_ioc_started_THEN_state_machine_initialized(self):
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:STATE", self.ca.Alarms.NONE)
-
-    def test_WHEN_auto_pres_ctrl_disabled_THEN_statemachine_in_init(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:STATUS", self.ca.Alarms.NONE)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_less_or_eq_tempsp_THEN_pressure_sp_set_to_pres_sp_minimum(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 50])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 60)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 20)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 50)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 20, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 20, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_less_than_cutoff_THEN_pressure_set_to_constant(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 15])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 10)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 1)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 20)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 5)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 5, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 5, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_less_than_cutoff_but_actual_temp_above_cutoff_THEN_pressure_not_set_to_constant(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 50])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 10)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 1)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 15)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 5)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 40, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 40, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_less_than_cutoff_THEN_pressure_bounded_by_minimum(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 10])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 10)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 12)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 15)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 7)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 12, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 12, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_less_than_cutoff_THEN_pressure_bounded_by_maximum(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 15])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 10)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 17)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 15)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 20)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 17, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 17, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_greater_than_cutoff_THEN_pressure_calculated_according_to_pressure_law(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 11])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 10)
-        self.ca.set_pv_value("VTI_SPC:TEMP:SCALE", 2)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 60)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 20)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-
-        # Pre-calculation: pressure_setpoint_min + temp_scale * (temperature - temperature_setpoint) ** 2
-        # 5 + 2 * (15 - 10) ** 2 = 55
-
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 7, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 7, tolerance=0.01)
-
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_greater_than_cutoff_THEN_pressure_calculated_according_to_interpolation_func(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 300])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 2)
-        self.ca.set_pv_value("VTI_SPC:TEMP:SCALE", 2)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 60)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 20)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-
-        # Temp      Pressure Max
-        # 0         13
-        # 300       13
-
-        self.ca.assert_that_pv_is_number("VTI_SPC:PRESSURE:SP:MAX:LKUP", 13, tolerance=0.01)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 13, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 13, tolerance=0.01)
-
-    def _change_spc_pressure_lookup_table(self, file_name):
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.NSPE", file_name)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.NMET", self.ca.get_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.METH"))
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.NBDI", self.ca.get_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.BDIR"))
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.NTDI", self.ca.get_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.TDIR"))
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX:LKUP.INIT", 1)
-        self.ca.assert_that_pv_is("VTI_SPC:PRESSURE:SP:MAX:LKUP.ISTA", "Done")
-        self.ca.process_pv("VTI_SPC:PRESSURE:SP:MAX:LKUP")
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_temp_sp_greater_than_cutoff_THEN_pressure_calculated_according_to_complex_interpolation_func(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 250])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 2)
-        self.ca.set_pv_value("VTI_SPC:TEMP:SCALE", 2)
-        self.ca.set_pv_value("VTI_SPC:TEMP:CUTOFF", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 5)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 60)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:CONST", 20)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self._change_spc_pressure_lookup_table("2x1x2x.txt")
-
-        # Temp      Pressure Max
-        # 0         5
-        # 100       15
-        # 200       20
-        # 300       30
-
-        self.ca.process_pv("VTI_SPC:PRESSURE:SP:MAX:LKUP")
-        self.ca.assert_that_pv_is_number("VTI_SPC:PRESSURE:SP:MAX:LKUP", 25, tolerance=0.01)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 25, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 25, tolerance=0.01)
-
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 50])
-        self.ca.assert_that_pv_is_number("VTI_SPC:PRESSURE:SP:MAX:LKUP", 10, tolerance=0.01)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_sp_min_greater_than_sp_max_THEN_statemachine_stops(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:ERROR", 0)
-
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 10)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 20)
-
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:ERROR", self.ca.Alarms.NONE)
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "No errors")
-
-        # Set up the PVs
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 20)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 10)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "error_delay")
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:STATE", self.ca.Alarms.NONE)
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "Min Pressure > Max")
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:ERROR", self.ca.Alarms.INVALID)
-
-    @contextlib.contextmanager
-    def _disconnect_device(self):
-        self._lewis.backdoor_set_on_device("connected", False)
-        try:
-            yield
-        finally:
-            self._lewis.backdoor_set_on_device("connected", True)
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_device_disconnected_THEN_pressure_unchanged(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        temp_card_pv_prefix = get_card_pv_prefix(TEMP_CARDS[0])
-        pressure_card_pv_prefix = get_card_pv_prefix(PRESSURE_CARDS[0])
-
-        # Set up the PVs for the low temp loop and start the statemachine
-        self._lewis.backdoor_run_function_on_device(
-            "backdoor_set_channel_property", [TEMP_CARDS[0], "temperature", 50])
-        self.ca.set_pv_value(f"{temp_card_pv_prefix}:TEMP:SP", 60)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 20)
-        self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MAX", 50)
-        self.ca.set_pv_value(f"{pressure_card_pv_prefix}:PRESSURE:SP", 40)
-        self.ca.assert_that_pv_is_not_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 20, tolerance=0.01)
-        
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 20, tolerance=0.01)
-
-        with self._disconnect_device():
-            self.ca.set_pv_value("VTI_SPC:PRESSURE:SP:MIN", 21)
-
-            self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "Temp read failure", timeout=60)
-            self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 20, tolerance=0.01)
-            
-        self.ca.assert_that_pv_is_number(f"{pressure_card_pv_prefix}:PRESSURE:SP:RBV", 21, tolerance=0.01)
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "No errors")
-
-    @skip_if_recsim("Lewis backdoor not available in recsim")
-    def test_GIVEN_sm_on_WHEN_pressure_card_invalid_THEN_statemachine_stops(self):
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "Off")
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:STATE", "init")
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:ERROR", 0)
-
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:ERROR", self.ca.Alarms.NONE)
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "No errors")
-
-        # Set up the PVs        
-        self.ca.set_pv_value("VTI_SPC:PRESSURE_ID", "nodevice")
-        self.ca.set_pv_value("VTI_SPC:STATEMACHINE:STATUS", "On")
-
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:STATE", self.ca.Alarms.NONE)
-        self.ca.assert_that_pv_is("VTI_SPC:STATEMACHINE:ERROR", "Invalid pressure card")
-        self.ca.assert_that_pv_alarm_is("VTI_SPC:STATEMACHINE:ERROR", self.ca.Alarms.MINOR)
-
-        # Put the card back how it should be.
-        self.ca.set_pv_value("VTI_SPC:PRESSURE_ID", PRESSURE_CARDS[0])
-
-
-
